intel_poller: Polling based on strategy.
