+++ pkg/poller/loader.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/shashank-mugiwara/laughingtale/db"
 	"github.com/shashank-mugiwara/laughingtale/logger"
+	"github.com/shashank-mugiwara/laughingtale/pkg/factory"
 	"github.com/shashank-mugiwara/laughingtale/pkg/ingest"
 	"github.com/shashank-mugiwara/laughingtale/pkg/type_configs"
 )
@@ -34,13 +35,24 @@ func GetAllLoaderSourceConfigs() {
 	}
 
 	for _, scfg := range loaderScenarioConfigs {
-		PollData(scfg)
+		identifier := scfg.Identifier
+		for _, cfg := range scfg.SourceConfig {
+			go ProcessEachSourceConfig(identifier, cfg)
+		}
 	}
 }
 
-func PollData(loaderSourceConfigs type_configs.SourceConfigs) {
-	identifier := loaderSourceConfigs.Identifier
-	for _, scfg := range loaderSourceConfigs.SourceConfig {
-		go ingest.PollDataFromSourceAndIngestToTarget(identifier, scfg)
+func ProcessEachSourceConfig(identifier string, sourceConfig type_configs.SourceConfig) {
+	pollingStrategy := sourceConfig.PollerConfig.PollingStrategy
+	poller, err := factory.GetStrategyFactory(pollingStrategy)
+	if err != nil {
+		logger.GetLaughingTaleLogger().Error(err.Error())
 	}
+
+	resultList, resultErr := poller.Poll(identifier, sourceConfig)
+	if resultErr != nil {
+		logger.GetLaughingTaleLogger().Error(resultErr.Error())
+	}
+
+	ingest.IngestPostgresToMongo(identifier, sourceConfig, resultList)
 }
