--- pkg/poller/strategy/concrete.go
@@ -1,6 +1,0 @@
-package poller_strategy
-
-type PollerStrategy struct {
-	WhereQueryPrefix         string
-	PollerFrequencyInSeconds int
-}
