--- pkg/poller/strategy/simple.go
@@ -1,20 +1,0 @@
-package poller_strategy
-
-import "database/sql"
-
-type SimpleStrategy struct {
-	PollerStrategy
-}
-
-func newSimpleStrategyPoller() IPollerStrategy {
-	return &SimpleStrategy{
-		PollerStrategy: PollerStrategy{
-			WhereQueryPrefix:         "",
-			PollerFrequencyInSeconds: 60,
-		},
-	}
-}
-
-func (simpleStrategy *SimpleStrategy) Poll() *sql.Rows {
-	return nil
-}
