--- pkg/poller/strategy/simple_incremental.go
@@ -1,1 +1,0 @@
-package poller_strategy
