--- pkg/poller/strategy/strategy.go
@@ -1,7 +1,0 @@
-package poller_strategy
-
-import "database/sql"
-
-type IPollerStrategy interface {
-	Poll() *sql.Rows
-}
