+++ pkg/strategy/impl_simple_incremental.go
@@ -1,0 +1,1 @@
+package poller_strategy
