+++ pkg/utils/db_utils.go
@@ -1,0 +1,104 @@
+package utils
+
+import (
+	"database/sql"
+
+	"github.com/shashank-mugiwara/laughingtale/db"
+	"github.com/shashank-mugiwara/laughingtale/logger"
+	"github.com/shashank-mugiwara/laughingtale/pkg/type_configs"
+)
+
+func ParseSqlRows(rows *sql.Rows) ([]interface{}, error) {
+	columnTypes, err := rows.ColumnTypes()
+
+	if err != nil {
+		return nil, err
+	}
+
+	count := len(columnTypes)
+	finalRows := []interface{}{}
+
+	for rows.Next() {
+
+		scanArgs := make([]interface{}, count)
+
+		for i, v := range columnTypes {
+
+			switch v.DatabaseTypeName() {
+			case "VARCHAR", "TEXT", "UUID", "TIMESTAMP":
+				scanArgs[i] = new(sql.NullString)
+				break
+			case "BOOL":
+				scanArgs[i] = new(sql.NullBool)
+				break
+			case "INT4":
+				scanArgs[i] = new(sql.NullInt64)
+				break
+			default:
+				scanArgs[i] = new(sql.NullString)
+			}
+		}
+
+		err := rows.Scan(scanArgs...)
+
+		if err != nil {
+			return nil, err
+		}
+
+		masterData := map[string]interface{}{}
+
+		for i, v := range columnTypes {
+
+			if z, ok := (scanArgs[i]).(*sql.NullBool); ok {
+				masterData[v.Name()] = z.Bool
+				continue
+			}
+
+			if z, ok := (scanArgs[i]).(*sql.NullString); ok {
+				masterData[v.Name()] = z.String
+				continue
+			}
+
+			if z, ok := (scanArgs[i]).(*sql.NullInt64); ok {
+				masterData[v.Name()] = z.Int64
+				continue
+			}
+
+			if z, ok := (scanArgs[i]).(*sql.NullFloat64); ok {
+				masterData[v.Name()] = z.Float64
+				continue
+			}
+
+			if z, ok := (scanArgs[i]).(*sql.NullInt32); ok {
+				masterData[v.Name()] = z.Int32
+				continue
+			}
+
+			masterData[v.Name()] = scanArgs[i]
+		}
+
+		finalRows = append(finalRows, masterData)
+	}
+
+	return finalRows, err
+}
+
+type RecordCountResult struct {
+	RecCount int `json:"rec_count"`
+}
+
+func ListTables(schema string) []string {
+	var tables []string
+	if err := db.GetlaughingtaleDb().Table("information_schema.tables").Where("table_schema = ?", "public").Pluck("table_name", &tables).Error; err != nil {
+		panic(err)
+	}
+	return tables
+}
+
+func GetRecordCount(sourceConfig type_configs.SourceConfig) int {
+	query := "SELECT COUNT(" + sourceConfig.PrimaryKey + ") as rec_count FROM " + sourceConfig.DbSchema + "." + sourceConfig.TableName + " WHERE " + sourceConfig.FilterConfig.WhereQuery
+	recordNumberResult := RecordCountResult{RecCount: 0}
+	db.GetlaughingtaleDb().Raw(query).Scan(&recordNumberResult)
+	logger.GetLaughingTaleLogger().Info("Found ", recordNumberResult.RecCount, " number of records from table: ", sourceConfig.TableName)
+	return recordNumberResult.RecCount
+}
